requirements.txt
# Core Deep Learning Framework
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Computer Vision Libraries
opencv-python>=4.8.0
ultralytics>=8.0.0
supervision>=0.15.0

# Multi-Object Tracking
deep-sort-realtime>=1.3.0
filterpy>=1.4.5

# Data Processing and Augmentation
numpy>=1.21.0
pandas>=1.5.0
albumentations>=1.3.0
Pillow>=9.0.0

# Visualization and Plotting
matplotlib>=3.6.0
seaborn>=0.12.0
plotly>=5.0.0

# Machine Learning Utilities
scikit-learn>=1.3.0
scipy>=1.9.0

# Dataset Management
roboflow>=1.1.0

# Web Interface
gradio>=3.40.0
streamlit>=1.25.0

# Development and Testing
pytest>=7.0.0
pytest-cov>=4.0.0
black>=23.0.0
flake8>=6.0.0

# Utilities
tqdm>=4.64.0
pyyaml>=6.0
python-dotenv>=1.0.0
requests>=2.28.0

# Optional: For enhanced functionality
timm>=0.9.0  # Additional pre-trained models
wandb>=0.15.0  # Experiment tracking
tensorboard>=2.13.0  # Training visualization

# Platform-specific dependencies
# For macOS with Apple Silicon
# --extra-index-url https://download.pytorch.org/whl/cpu

# For CUDA support (uncomment if using NVIDIA GPU)
# --extra-index-url https://download.pytorch.org/whl/cu118